requirements.txt
bleach==6.2.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kaggle==1.7.4.5
kagglehub==0.3.12
lightgbm==4.6.0
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
nltk==3.9.1
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
Panda3D==1.10.15
panda3d-gltf==1.3.0
panda3d-simplepbr==0.13.1
pandas==2.3.0
pillow==11.3.0
protobuf==6.31.1
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
scikit-learn==1.7.0
scipy==1.16.0
screeninfo==0.8.1
setuptools==80.9.0
six==1.17.0
sswg==4.3.0
text-unidecode==1.3
textblob==0.19.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
ursina==8.0.1
webencodings==0.5.1
Werkzeug==3.1.3
